refactor: extract runtime admin menu concern
